update express routing to split api and serve react bundle
