+++ .circleci/config.yml
@@ -11,6 +11,7 @@ jobs:
           - v1-dependencies-{{ checksum "package.json" }}
           - v1-dependencies-
       - run: npm install
+      - run: npm run build
       - save_cache:
           paths:
             - node_modules
