+++ babel.js
@@ -1,4 +1,4 @@
 // FIXME webpack
 require('@babel/register')
 require('@babel/polyfill')
-require('./server/app/api')
+require('./server/app/')
