+++ lib/logger.js
@@ -8,7 +8,7 @@ import {
   PAPERTRAIL_HOST,
   PAPERTRAIL_PORT,
   PAPERTRAIL_LEVEL
-} from '../server/app/config'
+} from './config'
 
 const { formatter, timestamp } = consoleFormatter()
