+++ lib/mail.js
@@ -1,10 +1,9 @@
-import postmark from 'postmark'
 import mailgun from 'mailgun-js'
 
 import log from '../lib/logger'
 import {
-  APP_DOMAIN, ENVIRONMENT, MAIL_FROM, POSTMARK_KEY, MAILGUN_KEY
-} from '../server/app/config'
+  APP_DOMAIN, ENVIRONMENT, MAIL_FROM, MAILGUN_KEY
+} from './config'
 
 export default function sendMail({ to, subject, text }) {
   log.info('sendMail attempt', { to, subject, text })
@@ -19,20 +18,5 @@ export default function sendMail({ to, subject, text }) {
     }
 
     mailer.messages().send(email, (err, res) => log.info('sendMail', err, res))
-  }
-}
-
-export const sendPostmarkMail = ({ To, Subject, TextBody }) => {
-  if (ENVIRONMENT === 'production') {
-    const mailer = new postmark.Client(POSTMARK_KEY)
-
-    let email = {
-      From: MAIL_FROM,
-      To,
-      Subject,
-      TextBody,
-    }
-
-    mailer.sendEmail(email, (e, res) => log.info('sendMail', email, e, res))
   }
 }
