+++ package.json
@@ -10,8 +10,8 @@
   "scripts": {
     "eslint": "eslint ./",
     "eslint:fix": "eslint ./ --fix",
-    "build": "webpack --config .webpack/config.js",
-    "start:dev:webpack": "webpack-dev-server --config .webpack/config.js",
+    "build": "webpack --config .webpack/config.js --mode=production",
+    "start:dev:webpack": "webpack-dev-server --config .webpack/config.js --watch --mode=development",
     "start": "node babel.js",
     "start:dev": "nodemon babel.js",
     "start:dev:debug": "DEBUG=sequelize* nodemon babel.js",
@@ -24,7 +24,6 @@
   },
   "dependencies": {
     "express": "^4.16.3",
-    "express-handlebars": "^3.0.0",
     "express-subdomain-handler": "^0.1.0",
     "express-useragent": "^1.0.12",
     "ipapi.co": "^0.2.3",
@@ -34,11 +33,8 @@
     "moment-business-days": "^1.0.6",
     "moment-timezone": "^0.5.16",
     "mysql": "^2.15.0",
-    "node-sass-middleware": "^0.11.0",
     "pg": "^7.4.2",
-    "pinkie-promise": "^2.0.1",
     "polished": "^2.3.1",
-    "postmark": "^1.6.1",
     "react": "^16.6.3",
     "react-dom": "^16.6.3",
     "react-router-dom": "^4.3.1",
@@ -77,7 +73,6 @@
     "eslint-plugin-promise": "^3.7.0",
     "eslint-plugin-standard": "^3.1.0",
     "fork-ts-checker-webpack-plugin": "^0.5.0",
-    "given2": "^2.1.5",
     "html-webpack-plugin": "^3.2.0",
     "mocha": "^5.1.1",
     "mocha-junit-reporter": "^1.17.0",
