+++ server/app/api.js
@@ -1,46 +1,52 @@
-import app from './express'
-import { ALLOW_ADMIN_ACTIONS, ENVIRONMENT } from '../app/config'
+import { Router } from 'express';
+
+import log from '../../lib/logger'
+import { ALLOW_ADMIN_ACTIONS, ENVIRONMENT } from '../../lib/config'
 import { Promises } from '../models/'
 import { seed, importJson } from '../db/seed'
 import cache from '../db/cache'
 
-import FrontendApi from '../api/v1/frontend'
-import PromiseApi from '../api/v1/promises'
-import UserApi from '../api/v1/users'
+import PromiseApi from '../api/v1/promise'
+import UserApi from '../api/v1/user'
 
-// TODO: https://github.com/Vincit/objection.js/tree/master/examples/express-ts
-
-// Client-side/browser API
-app.use('/', FrontendApi)
+const api = Router();
 
 // REST endpoints
-app.use('/api/v1', PromiseApi)
-app.use('/api/v1', UserApi)
+api.use('/promise', PromiseApi)
+api.use('/user', UserApi)
 
 // Utils
 
 // calculate and store reliability for each user
-app.get('/cache', (req, resp) => {
+api.get('/cache', (req, resp) => {
   cache()
   resp.redirect('/')
 })
 
 if (ENVIRONMENT !== 'production' || ALLOW_ADMIN_ACTIONS) {
   // insert promises.json into db
-  app.get('/import', (req, resp) => {
+  api.get('/import', (req, resp) => {
     importJson()
     resp.redirect('/')
   })
 
   // drop db and repopulate
-  app.get('/reset', (req, resp) => {
+  api.get('/reset', (req, resp) => {
     seed()
     resp.redirect('/')
   })
 
   // removes all entries from the promises table
-  app.get('/empty', (req, resp) => {
+  api.get('/empty', (req, resp) => {
     Promises.destroy({ where: {} })
     resp.redirect('/')
   })
 }
+
+// // catch-all
+api.get('*', (req, res) => {
+  log.info('render 401', req.originalUrl)
+  res.send(401)
+})
+
+export default api;
