+++ server/app/index.js
@@ -1,9 +1,10 @@
 import express from 'express'
 import useragent from 'express-useragent'
+import path from 'path';
 
-import { PORT } from './config'
+import { PORT } from '../../lib/config'
 import log from '../../lib/logger'
-
+import apiRouter from './api';
 
 const app = express()
 
@@ -16,14 +17,18 @@ app.use(express.urlencoded({
   extended: false
 }))
 
+app.use(express.static(path.join(__dirname, '../../build')));
+
 app.listen(PORT, () => {
   log.info(`The commits.to app is running on port ${PORT}`)
 })
 
+app.use('/api/v1', apiRouter);
+
 // catch-all
-// app.get('*', (req, res) => {
-//   log.info('render 401', req.originalUrl)
-//   res.send(401)
-// })
+app.get('*', (req, res) => {
+  log.info('render front-end', req.originalUrl)
+  res.sendFile(path.join(`${__dirname}/../../build/index.html`))
+})
 
 export default app
